transport: consolidate common code between virtiofs and fusedev
Consolidate common between virtiofs and fusedev, it's a preparation
to avoid mutual exclusive between virtiofs and fusedev.
